[cross] CalcListPos returns float, not double
The -1.0f sentinel and final result are loaded/returned as single precision
(lfs FLOAT_8033299C), indicating CalcListPos returns float. Update the
declaration in p_menu.h to match; callers already implicitly widen to double.

+++ include/ffcc/p_menu.h
@@ -289,7 +289,7 @@ public:
     void DrawBar(float, float, float, TEX, float);
     void DrawWindow(float, float, float, float, TEX, float);
     void DrawSingleIcon(int, int, int, float, int, float);
-    double CalcListPos(int, int, int);
+    float CalcListPos(int, int, int);
     void DrawListPosMark(float, float, float);
     void DrawSingLife();
     float CalcCenteringPos(char*, CFont*);
